Partie 4 tableaux finalisée

+++ 01 - Algo/03-Boucles/Exo_3_7_Barnabe/Program_3_7.cs
@@ -11,7 +11,7 @@
 
 
             Console.WriteLine("Veuillez entrer une somme de départ : ");
-            sommeDepart = int.Parse(Console.ReadLine());
+            sommeDepart = Double.Parse(Console.ReadLine());
             
 
             do
